fc_test/requirements.txt
featurecloud
wheel
scikit-learn
Pillow
torch
torchvision
